refactor: remove network file
