+++ lib/artnet.js
@@ -1,7 +1,7 @@
+import os from 'node:os';
 import dgram from 'node:dgram';
 import { Buffer } from 'node:buffer';
 import Util from './util.js';
-import Network from './network.js';
 
 const NODE_NAME = 'HueDMX';
 const PORT = 6454;
@@ -10,6 +10,14 @@ const OPCODE_OUTPUT = 0x50_00; // ArtDMX data packet
 const OPCODE_POLL = 0x20_00; // ArtPoll packet
 const OPCODE_POLLREPLY = 0x21_00; // ArtPollReply packet
 const MIN_PROTOCOL_VERSION = 14;
+
+export function getNetworkInterface() {
+  const interfaces = os.networkInterfaces();
+  return Object.keys(interfaces)
+    .map((n) => interfaces[n])
+    .map((n) => n.find((i) => !i.internal && i.family === 'IPv4'))
+    .find(Boolean);
+}
 
 export function listenArtNet(address, dmxHandler) {
   // Set up the socket
@@ -40,7 +48,7 @@ export function listenArtNet(address, dmxHandler) {
       } else if (opcode === OPCODE_POLL) {
         // Send ArtPollReply
         const buffer = Buffer.alloc(238);
-        const itf = Network.interface();
+        const itf = getNetworkInterface();
         const ip = itf.address.split('.').map((s) => Number.parseInt(s, 10));
         const mac = itf.mac.split(':').map((s) => Number.parseInt(s, 16));
         buffer.write(HEADER, 0, 8);
